+++ lib/modules/core/core_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: core
-// Updated at: 2026-06-05 21:07:45
-// Commit: feat(core): add bi-directional event bus handler for core
+// Updated at: 2026-06-06 17:46:50
+// Commit: fix(core): resolve race condition during core initialization
 
 class coreComponentState {
-  final String commitRef = '1789';
-  final String lastMessage = 'feat(core): add bi-directional event bus handler for core';
-  final DateTime timestamp = DateTime.parse('2026-06-05T21:07:45+05:00');
+  final String commitRef = '1811';
+  final String lastMessage = 'fix(core): resolve race condition during core initialization';
+  final DateTime timestamp = DateTime.parse('2026-06-06T17:46:50+05:00');
 }